refactor: :art: api 디렉토리 구조 변경, 헤더 속성 이름 변경
