+++ src/app.ts
@@ -9,7 +9,7 @@ export default async function app(fastify: FastifyInstance) {
   setDecorate(fastify);
   setMiddleware(fastify);
 
-  fastify.register(routeV1, { prefix: '/v1' });
+  fastify.register(routeV1, { prefix: '/users/v1' });
 }
 
 function setErrorHandler(fastify: FastifyInstance) {
@@ -25,22 +25,22 @@ function setErrorHandler(fastify: FastifyInstance) {
 
 function setMiddleware(fastify: FastifyInstance) {
   fastify.addHook('onRequest', (request, reply, done) => {
-    const authorized = request.headers['x-authorized'];
+    const authenticated = request.headers['x-authenticated'];
     const userId = request.headers['x-user-id'];
     
-    if (authorized === undefined || Array.isArray(authorized)) {
+    if (authenticated === undefined || Array.isArray(authenticated)) {
       done();
     }
     if (userId === undefined || Array.isArray(userId)) {
       done();
     }
-    if (isNaN(Number(userId)) === true) {
+    if (isNaN(Number(userId))) {
       throw new UnAuthorizedException('user id is not a number');
     }
 
-    if (authorized === 'true') {
-      request.authorized = true;
-      request.myId = parseInt(userId as string , 10);
+    if (authenticated === 'true') {
+      request.authenticated = true;
+      request.userId = parseInt(userId as string , 10);
     }
     done();
   });
@@ -48,8 +48,8 @@ function setMiddleware(fastify: FastifyInstance) {
 
 function setDecorate(fastify: FastifyInstance) {
   fastify.decorate('authenticate', async (request: FastifyRequest, reply: FastifyReply) => {
-    console.log(request.authorized);
-    if (!request.authorized) {
+    console.log(request.authenticated);
+    if (!request.authenticated) {
       reply.code(401).send({
         status: STATUS.ERROR,
         message: 'Unauthorized',
