+++ src/v1/apis/users.controller.ts
@@ -7,8 +7,8 @@ export default class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   findUser = async (request: FastifyRequest, reply: FastifyReply) => {
-    request.log.info(request.authorized, "authorized");
-    request.log.info(request.myId, "myid");
+    request.log.info(request.authenticated, "authenticated");
+    request.log.info(request.userId, "userId");
     const params = getUserParamsSchema.parse(request.params);
     const result = await this.usersService.findUser(params.id);
     reply.code(200).send(result);
