+++ src/v1/apis/users.schema.ts
@@ -1,4 +1,4 @@
-import { createResponseSchema } from '../../common/schema/core.schema.js';
+import { createResponseSchema } from '../common/schema/core.schema.js';
 import { z } from 'zod';
 
 export const FindUserResponseSchema = createResponseSchema(
