+++ src/v1/common/types/fastify.d.ts
@@ -7,11 +7,11 @@ declare module 'fastify' {
 
   interface FastifyRequest {
     headers: {
-      'X-Authorized': string | undefined | string[];
+      'X-Authenticated': string | undefined | string[];
       'X-User-Id': string | undefined;
     };
 
-    authorized: boolean;
-    myId?: number;
+    authenticated: boolean;
+    userId?: number;
   }
 }
